Create a symlink to every alias's node binary.
When an alias is created, a directory named <alias>_bin is also created
in the alias directory, and a symlink pointing to the target's node
binary is created within that directory. This allows you to, for
example, make that alias's node the default node binary for all users,
by placing a symlink to the alias's symlink in, say, /usr/bin.

This is most handy for ensuring that node scripts which must be run as
root, such as init scripts that change their uids, will have access to a
node binary.