Fix `nvm ls` with multiple installed versions
The `-a` flag for basename, dropped in commit 335456d, was needed
to instruct it to not ignore the arguments after the first one.

Go back to using `\ls` as done before commit 42915fc, but using basename
and xargs to avoid changing the current directory.

With this the 'Running "nvm ls" should display all installed versions.'
test passes again.