Introduce nvm function for fish
This is a wrapper around nvm. It's main function is to check and promote environment variables changes from nvm to fish.

A little more is done to setup the environment in case that hasn't been done yet:

* install a stable node version if it doesn't exist
* set the default to stable alias if that hasn't been done yet

Due to those actions, the first execution of `nvm` might take some time.