Only use NVM on interactive login shells
i.e.  avoid using it on non-interactive or non-login shells

There is a big problem with the usage of `.bashrc` for outputing escape sequences (as-is), because it can corrupt the communication of other tools that are doing non-login SSH sessions (yes, some even use an interactive shell for non-interactive use). One of such problems we have with Ansible (see ansible#14282), but the same could happen to other tools.

So my prefered solution is to avoid using `.bashrc` to ensure we have a login-shell. But you may also want to ensure that you have an interactive shell before appending your own stuff. So I added that snippet as well. I am not sure if this would work on all distributions, so please test !

References:
https://www.gnu.org/software/bash/manual/html_node/Is-this-Shell-Interactive_003f.html
http://superuser.com/questions/183870/difference-between-bashrc-and-bash-profile