Issue #1015 resolved.
This strips trailing whitespace from the version as read from .nvmrc before attempting to use it.
Added a test for it, but not too familiar with this type of testing, so not sure how good of a test
it is.

Also had some trouble running the tests in general, even before I made any changes, I had some issues.
So somebody should definitely verify this all before merging.