example to automatically run 'nvm use' in Bash
I added an example to automatically `nvm use` when changing directories in a Bash shell, similar to the one for Zsh. The function is the same as the Zsh example, but the method of implementing overrides the `cd` command in bash, and then calls the built-in `cd` before calling the `load-nvmrc` function. 

I've been using this on OSX for a bit now, and it seems to be working well in my basic testing.