[Fix] checksums: remove libressl command support
This was an "oops" on my part when I was testing changes for my work
that was later adapted & merged in https://github.com/creationix/nvm/commit/f1bca106a8f5fc6a4b88b5ef30603cf31ea531f1.
LibreSSL is never installed as a `libressl` as it is intended to be a
replacement for `openssl` as much as possible, and consequently the
executable is called `openssl`.