Adds support for OpenBSD.
In order to download and unpack node source files under OpenBSD it is
necessary to run the extraction into two steps. Since `tar` and `xz` are
not integrated, the `-J` option is not available on OpenBSD default
`tar` tool.

This commit adds a test in order to figure this out, and handles the
proper file extraction under OpenBSD. The previous behaviour for other
Operating Systems should be untouched (more testing is welcome).

This commit does not yet handles the next issue, which is to compile
certain versions of node under OpenBSD. This issue is a bit more
delicated, since OpenBSD uses LibreSSL, and node support for it is still
not complete.

 - Adds check for tar's support for J option