Make nvm visible to shell children
Shell functions are not automatically exported when declared, which
turned out to be problematic when sourcing the `nvm.sh` file from a
login shell. Non-login shells, which are children of the login shell,
do not see the `nvm` function when it's sourced from `~/.profile` and
other similar profile files.

Take the following code as an example (using bash):

    $ foo() { echo 1; } # in login shell
    $ foo
    1
    $ bash # in non-login shell now
    $ foo
    bash: foo: command not found

`foo` is not seen by the second bash instance, even though it has been
declared by its parent process.

One of the solutions is to export `foo` right after declaring it by
doing `export -f NAME`:

    $ foo() { echo 1; } # in login shell
    $ export -f foo
    $ bash # in non-login shell now
    $ foo
    1

This commit does the same to the `nvm` function and its dependencies.