Dockerfile: Remove mirror site setting; it broke the build
The ubuntu.cs.utah.edu mirror appears to be obsolete; trying to use it
produces error messages along the lines of:

    Ign:2 https://ubuntu.cs.utah.edu/ubuntu jammy-updates InRelease
    Err:2 https://ubuntu.cs.utah.edu/ubuntu jammy-updates InRelease
      Certificate verification failed: The certificate is NOT trusted. The certificate issuer is unknown.  Could not handshake: Error in the certificate verification. [IP: 155.98.64.89 443]
    W: http://ubuntu.cs.utah.edu/ubuntu/dists/jammy-updates/InRelease: No system certificates available. Try installing ca-certificates.

These messages don't make it completely obvious what the problem is, and
it's also a rather obscure setting; you have to carefully read the
Dockerfile to find it.

Due to issues like this, and changing mirrors over time, it's best
just to drop attempts to use a specific mirror and use whatever the
default you get for archive.ubuntu.com etc.

The package downloads generally happen only once per host, anyway,
since Docker will keep not only the final image, but also the layers
used to generate it: if the final image is deleted or anything in the
Dockerfile after the OS package updates is changed, these cached
layers will be used.